working on fixing projects page
