+++ src/components/Home.module.css
@@ -1,5 +1,5 @@
 .uzi {
-    width: 75%;
+    width: 50%;
     border-radius: 150px;
     position: relative;
     transform: translate(10%, 0%);
