+++ src/components/NavBar.module.css
@@ -2,7 +2,7 @@
     display: flex;
     align-items: center;
     justify-content: space-between;
-    background-color: white;
+    background-color: rgb(215, 215, 215);
 }
 
 .pages {
@@ -39,11 +39,11 @@
 
 .uzi {
     width: 50px;
-    border-radius: 8px;
+    border-radius: 50px;
 }
 
 .name {
-    font-family: var(--bold-font);
+    font-family: var(--heading-font);
     font-weight: 900;
     font-style: normal;
     font-size: 48px;
