+++ src/components/Project.module.css
@@ -12,11 +12,14 @@ li {
 }
 
 .content {
-    width: 100%;
-    max-width: auto;
     display: flex;
-    flex-direction: column;
-    align-items: center;
+    flex-wrap: wrap;
+    justify-content: space-around;
+}
+
+.content > * {
+    width: 300px;
+    margin: 16px;
 }
 
 img {
