+++ src/components/ProjectCardData.js
@@ -1,0 +1,29 @@
+const ProjectCardData = [
+  {
+    id: 1,
+    title: "Hyzerberg Technologies",
+    description: "Creating a disc golf simulator for senior Capstone project",
+  },
+  {
+    id: 2,
+    title: "Smart Robot",
+    description: "",
+  },
+  {
+    id: 3,
+    title: "Lionize",
+    description: "",
+  },
+  {
+    id: 4,
+    title: "Ultrasonic Eyes",
+    description: "",
+  },
+  {
+    id: 5,
+    title: "Research",
+    description: "",
+  },
+];
+
+export default ProjectCardData;
